% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisParameters.R
\name{analysisParameters}
\alias{analysisParameters}
\title{module analysisParameter server function}
\usage{
analysisParameters(input, output, session, data, predefined.parameters)
}
\description{
module analysisParameter server function
}
