% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bins.position}
\alias{bins.position}
\title{Annotate all the peaks with their corresponding bin}
\usage{
bins.position(bins, peaks, ladder.samples)
}
\arguments{
\item{bins}{result of the read.bin.file function}

\item{peaks}{data.table produced by peaks.to.marker}

\item{ladder.samples}{vector of ladder samples}
}
\value{
a data table containing the position of the annotated bins
}
\description{
Annotate all the peaks with their corresponding bin
}
