% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExporter.R
\name{dataExporter}
\alias{dataExporter}
\title{module dataExporter server function}
\usage{
dataExporter(input, output, session, dataExporterFilters, annotated.peaks,
  fsa.data, colors, markers, markersList, seqdates)
}
\description{
module dataExporter server function
}
