% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExporterFilter.R
\name{dataExporterFilter}
\alias{dataExporterFilter}
\title{module dataExporterFilter server function}
\usage{
dataExporterFilter(input, output, session, fsa.data, selected.scale,
  exportTable)
}
\description{
module dataExporterFilter server function
}
