% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPeaks.R
\name{exportPeaks}
\alias{exportPeaks}
\title{module exportPeaks server function}
\usage{
exportPeaks(input, output, session, exportPeaksTable, colors, markers,
  markersList, seqdates, selected.sample)
}
\description{
module exportPeaks server function
}
