% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightSelector.R
\name{heightSelector}
\alias{heightSelector}
\title{module heightSelector server function}
\usage{
heightSelector(input, output, session, data, selected.dyes)
}
\description{
module heightSelector server function
}
