% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegressionViewer.R
\name{linearRegressionViewer}
\alias{linearRegressionViewer}
\title{module linearRegressionViewer  sever function}
\usage{
linearRegressionViewer(input, output, session, fsa.data)
}
\description{
module linearRegressionViewer  sever function
}
