% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegressionViewer.R
\name{linearRegressionViewerUI}
\alias{linearRegressionViewerUI}
\title{module linearRegressionViewer  UI function}
\usage{
linearRegressionViewerUI(id)
}
\description{
module linearRegressionViewer  UI function
}
