% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSaveParams.R
\name{loadSaveParams}
\alias{loadSaveParams}
\title{module loadSaveParams server function}
\usage{
loadSaveParams(input, output, session, data, global.parameters)
}
\description{
module loadSaveParams server function
}
