% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{markedpeaks.to.real.bins}
\alias{markedpeaks.to.real.bins}
\title{This function annotates the detected peaks in the ladder sample(s)  and the peaks that are annotated for this ladder.}
\usage{
markedpeaks.to.real.bins(bins, peaks, ladder.samples)
}
\arguments{
\item{bins}{result of the read.bin.file function}

\item{peaks}{data.table with the detected peaks}

\item{ladder.samples}{a vector containing the names of the ladder samples in the peaks data.table}
}
\value{
an annotated set of bins in a list
}
\description{
This function annotates the detected peaks in the ladder sample(s)  and the peaks that are annotated for this ladder.
}
