% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleExperimentsRefresher.R
\name{multipleExperimentsRefresher}
\alias{multipleExperimentsRefresher}
\title{module multipleExperimentsRefresher server function}
\usage{
multipleExperimentsRefresher(input, output, session, data)
}
\description{
module multipleExperimentsRefresher server function
}
