% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleExperimentsRefresher.R
\name{multipleExperimentsRefresherUI}
\alias{multipleExperimentsRefresherUI}
\title{module multipleExperimentsRefresher UI function}
\usage{
multipleExperimentsRefresherUI(id)
}
\description{
module multipleExperimentsRefresher UI function
}
