% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleExperimentViewer.R
\name{multipleExperimentViewer}
\alias{multipleExperimentViewer}
\title{module multipleExperimentViewer server function}
\usage{
multipleExperimentViewer(input, output, session, fsa.data, colors,
  selected.samples, selected.height, selected.width, selected.scale,
  selected.dyes, nbSamplesPerPage, pageNb, above.samples, pageRefreshed)
}
\description{
module multipleExperimentViewer server function
}
