% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleExperimentViewer.R
\name{multipleExperimentViewerUI}
\alias{multipleExperimentViewerUI}
\title{module multipleExperimentViewer UI function}
\usage{
multipleExperimentViewerUI(id)
}
\description{
module multipleExperimentViewer UI function
}
