% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleViewerPageSelector.R
\name{multipleViewerPageSelector}
\alias{multipleViewerPageSelector}
\title{module multipleViewerPageSelectorUI server function}
\usage{
multipleViewerPageSelector(input, output, session, data, selected.samples,
  parameters)
}
\description{
module multipleViewerPageSelectorUI server function
}
