% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleViewerPageSelector.R
\name{multipleViewerPageSelectorUI}
\alias{multipleViewerPageSelectorUI}
\title{module multipleViewerPageSelectorUI UI function}
\usage{
multipleViewerPageSelectorUI(id)
}
\description{
module multipleViewerPageSelectorUI UI function
}
