% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleViewerSampleSelector.R
\name{multipleViewersampleSelector}
\alias{multipleViewersampleSelector}
\title{module multipleViewersampleSelector server function}
\usage{
multipleViewersampleSelector(input, output, session, data)
}
\description{
module multipleViewersampleSelector server function
}
