% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleViewerSampleSelector.R
\name{multipleViewersampleSelectorUI}
\alias{multipleViewersampleSelectorUI}
\title{module multipleViewersampleSelector UI function}
\usage{
multipleViewersampleSelectorUI(id)
}
\description{
module multipleViewersampleSelector UI function
}
