% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my.read.fsa}
\alias{my.read.fsa}
\title{Wrapper around seqinr::read.abif to generate a list with the content of a set of FSA files}
\usage{
my.read.fsa(files, dyes.default = c(Dye1 = "6-FAM", Dye2 = "HEX", Dye3 =
  "NED", Dye4 = "ROX"))
}
\arguments{
\item{files}{vector of FSA file(s)}

\item{dyes.default}{vector containing the id of the dye name (default is c(Dye1 = '6-FAM', Dye2 = 'HEX', Dye3 = 'NED', Dye4 = 'ROX'))}
}
\value{
a list containing abif.data : a data.table containing the raw data of the fsa file; dyes : of vector containing the names of dyes; expdate : a named vector containing the experiment date
}
\description{
Wrapper around seqinr::read.abif to generate a list with the content of a set of FSA files
}
