% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnalyzer.R
\name{peakAnalyzer}
\alias{peakAnalyzer}
\title{module peakAnalyzer server function}
\usage{
peakAnalyzer(input, output, session, selected.scale, parameters, data,
  global.parameters, predefined.parameters)
}
\description{
module peakAnalyzer server function
}
