% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{peaks.to.markers}
\alias{peaks.to.markers}
\title{Given the (rescaled) intensities, detects the peaks and annotate them with the system they are supposed to belong to.}
\usage{
peaks.to.markers(fsa.data, minpeakheights, removeStutters)
}
\arguments{
\item{fsa.data}{}

\item{minpeakheights}{do not consider peaks that have lower intensities than minpeakheights. Must be a list[[samplename]][[dyename]]}

\item{boolean}{removes stutter peaks}
}
\value{
for each sample and each probe a set of peaks (in a data.table) together with the system they are interesting with.
}
\description{
Given the (rescaled) intensities, detects the peaks and annotate them with the system they are supposed to belong to.
}
