% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read.bin.file}
\alias{read.bin.file}
\title{Convert a bin file (GeneMapper idx format) to a data.Table that can be used by}
\usage{
read.bin.file(bin.file)
}
\arguments{
\item{bin.file}{as used by GeneMapper that contains the position of the bins of a ladder sample.}
}
\value{
a data table containing for each peak the bin it is inferred to belong to
}
\description{
Convert a bin file (GeneMapper idx format) to a data.Table that can be used by
}
