% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelector.R
\name{sampleSelectorUI}
\alias{sampleSelectorUI}
\alias{sampleSelector}
\title{sampleSelector and sampleSelectorUI}
\usage{
sampleSelectorUI(id)

sampleSelector(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
