% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale.timeseries}
\alias{scale.timeseries}
\title{Based on a standard.dye and a ladder, the function convert the time into molecular weight (based on a linear regression model)}
\usage{
\method{scale}{timeseries}(fsa.raw.data, time = time, scales,
  ladder = "LIZ500", standard.dye = "LIZ", minpeakheights = NULL,
  removeOutlyers = TRUE, minDist = 10)
}
\arguments{
\item{fsa.raw.data}{data.table containing the itensity values for each dye and each sample}

\item{time}{column of fsa.raw.data data.table containing the values that should be scaled according to the scaling dye}

\item{scales}{list of scales into data.tables}

\item{ladder}{type of scaling system}

\item{standard.dye}{id of the dye containing the scaling}

\item{minpeakheights}{minimal intensity value to be considered as a peak}

\item{removeOutlyers}{deprecated, not used anymore}

\item{minDist}{minimal time distance between two peaks}
}
\value{
a data table containing the molecular size values (obtained by normalizing accoding to a given scale of a given dye) vs the intensity values
}
\description{
Based on a standard.dye and a ladder, the function convert the time into molecular weight (based on a linear regression model)
}
