% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleSelector.R
\name{scaleSelector}
\alias{scaleSelector}
\title{module scaleSelector server function}
\usage{
scaleSelector(input, output, session, data, scales, parameters,
  predefined.parameters)
}
\description{
module scaleSelector server function
}
