% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectedPeakDisplayer.R
\name{selectedPeakDisplayer}
\alias{selectedPeakDisplayer}
\title{module selectedPeakDisplayer  server function}
\usage{
selectedPeakDisplayer(input, output, session, selectedPeak)
}
\description{
module selectedPeakDisplayer  server function
}
