% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectedPeakDisplayer.R
\name{selectedPeakDisplayerUI}
\alias{selectedPeakDisplayerUI}
\title{module selectedPeakDisplayer  UI function}
\usage{
selectedPeakDisplayerUI(id)
}
\description{
module selectedPeakDisplayer  UI function
}
