% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleExperimentFiltersAndLayouts.R
\name{singleExperimentFiltersAndLayouts}
\alias{singleExperimentFiltersAndLayouts}
\title{module singleExperimentFiltersAndLayouts server function}
\usage{
singleExperimentFiltersAndLayouts(input, output, session, data)
}
\description{
module singleExperimentFiltersAndLayouts server function
}
