% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleExperimentFiltersAndLayouts.R
\name{singleExperimentFiltersAndLayoutsUI}
\alias{singleExperimentFiltersAndLayoutsUI}
\title{module singleExperimentFiltersAndLayouts UI function}
\usage{
singleExperimentFiltersAndLayoutsUI(id)
}
\description{
module singleExperimentFiltersAndLayouts UI function
}
