% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleExperimentPeakAnalyzer.R
\name{singleExperimentPeakAnalyzerUI}
\alias{singleExperimentPeakAnalyzerUI}
\title{module singleExperimentPeakAnalyzer UI function}
\usage{
singleExperimentPeakAnalyzerUI(id)
}
\description{
module singleExperimentPeakAnalyzer UI function
}
