% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleExperimentViewer.R
\name{singleExperimentViewer}
\alias{singleExperimentViewer}
\title{module singleExperimentViewer server function}
\usage{
singleExperimentViewer(input, output, session, fsa.data, colors,
  singleExperimentFilterDyes, singleExperimentFilterExp,
  singleExperimentYaxis, singleExperimentFilterSystem,
  singleExperimentSystemDyeSelector, allSystemsSameLine,
  minValueFilterThresholdField, minValueFilterThresholdButton,
  includeExcludeButton, ladder.sample)
}
\description{
module singleExperimentViewer server function
}
