% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleExperimentViewer.R
\name{singleExperimentViewerUI}
\alias{singleExperimentViewerUI}
\title{module singleExperimentViewer UI function}
\usage{
singleExperimentViewerUI(id)
}
\description{
module singleExperimentViewer UI function
}
