% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widthSelector.R
\name{widthSelector}
\alias{widthSelector}
\title{module widthSelector server function}
\usage{
widthSelector(input, output, session, data, selected.scale)
}
\description{
module widthSelector server function
}
